[ai-assisted] refactor(profile): introduce feature module
Why:
- Address issue #54 by piloting the React feature module structure on the profile domain.

What:
- Moved profile API, query keys, page, and password dialog under src/react/features/profile.
- Added a profile route module and updated AppRouter to consume it.
- Updated FullLayout to import the password dialog from the feature module.
- Updated CHANGELOG.md for issue #54.

Validation:
- npm run typecheck: pass
- npm run lint: pass with existing warnings only
- npm run build: pass; Vite reported existing large chunk warning
