+++ src/react/features/profile/routes.tsx
@@ -1,0 +1,6 @@
+import { Route } from "react-router-dom";
+import { MyProfilePage } from "@/react/features/profile/pages/MyProfilePage";
+
+export const profileRoutes = (
+  <Route path="profile" element={<MyProfilePage />} />
+);
